+++ components-api/src/test/scala/pl/touk/nussknacker/engine/api/definition/ParameterValidatorSpec.scala
@@ -3,8 +3,11 @@ package pl.touk.nussknacker.engine.api.definition
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks
-import pl.touk.nussknacker.engine.api.NodeId
+import org.springframework.expression.spel.standard.SpelExpressionParser
+import org.springframework.expression.spel.support.StandardEvaluationContext
+import pl.touk.nussknacker.engine.api.{Context, NodeId}
 import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.api.expression.{Expression => ApiExpression}
 
 import scala.jdk.CollectionConverters._
 
@@ -221,6 +224,46 @@ class ParameterValidatorSpec extends AnyFunSuite with TableDrivenPropertyChecks 
     ) { (expression, value, validator, expected) =>
       validator.isValid("dummy", Expression.spel(expression), value, None).isValid shouldBe expected
     }
+  }
+
+  test("ValidationExpressionParameterValidator") {
+    forAll(
+      Table(
+        ("validationExpression", "paramName", "inputExpression", "value", "isValid"),
+        ("#param > 10", "param", "", Some(null), true),
+        ("#param > 10", "param", "#input.foo", None, true),
+        ("#param > 10", "param", "-14", Some(-14), false),
+        (
+          "#param.toLowerCase() == \"left\" || #param.toLowerCase() == \"right\"",
+          "param",
+          "'lEfT'",
+          Some("lEfT"),
+          true
+        ),
+        ("#param.toLowerCase() == \"left\" || #param.toLowerCase() == \"right\"", "param", "'up'", Some("up"), false),
+        ("#param", "param", "'up'", Some("up"), false),
+        ("#fooBar", "param", "'up'", Some("up"), false),
+        ("#list.size() == 2 && #list[0] == 'foo'", "list", "{'foo', 'bar'}", Some(List("foo", "bar").asJava), true)
+      )
+    ) { (validationExpression, paramName, inputExpression, value, isValid) =>
+      ValidationExpressionParameterValidator(new TestSpelExpression(validationExpression), None)
+        .isValid(paramName, Expression.spel(inputExpression), value, None)(nodeId)
+        .isValid shouldBe isValid
+    }
+  }
+
+}
+
+private class TestSpelExpression(expression: String) extends ApiExpression {
+
+  override def language: String = "spel"
+
+  override def original: String = expression
+
+  override def evaluate[T](ctx: Context, globals: Map[String, Any]): T = {
+    val context = new StandardEvaluationContext()
+    ctx.variables.foreach { case (param, value) => context.setVariable(param, value) }
+    new SpelExpressionParser().parseRaw(expression).getValue(context).asInstanceOf[T]
   }
 
 }