+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/validation/UIProcessValidatorSpec.scala
@@ -2,7 +2,7 @@ package pl.touk.nussknacker.ui.validation
 
 import cats.data.{Validated, ValidatedNel}
 import com.typesafe.config.ConfigValueFactory.{fromAnyRef, fromIterable}
-import com.typesafe.config.{Config, ConfigFactory}
+import com.typesafe.config.{Config, ConfigFactory, ConfigValueFactory}
 import org.scalatest.Inside.inside
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
@@ -22,7 +22,11 @@ import pl.touk.nussknacker.engine.canonicalgraph.canonicalnode.{FlatNode, SplitN
 import pl.touk.nussknacker.engine.graph.EdgeType.{NextSwitch, SwitchDefault}
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.fragment.FragmentRef
-import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{FragmentClazzRef, FragmentParameter}
+import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{
+  FragmentClazzRef,
+  FragmentParameter,
+  ValidationExpression
+}
 import pl.touk.nussknacker.engine.graph.node._
 import pl.touk.nussknacker.engine.graph.service.ServiceRef
 import pl.touk.nussknacker.engine.graph.sink.SinkRef
@@ -768,6 +772,72 @@ class UIProcessValidatorSpec extends AnyFunSuite with Matchers {
             List(
               NodeValidationError("EmptyMandatoryParameter", _, _, Some("P1"), NodeValidationErrorType.SaveAllowed),
               NodeValidationError("EmptyMandatoryParameter", _, _, Some("P2"), NodeValidationErrorType.SaveAllowed)
+            )
+          ) =>
+    }
+  }
+
+  test("validates scenario with fragment parameters - with spel validation expression and valid value") {
+    val fragmentId = "fragment1"
+    val paramName  = "name"
+
+    val fragmentDefinition: CanonicalProcess =
+      createFragmentDefinition(
+        fragmentId,
+        List(
+          FragmentParameter(
+            paramName,
+            FragmentClazzRef[String],
+            Some(ValidationExpression(s"#$paramName.length() < 7"))
+          )
+        )
+      )
+    val processWithFragment =
+      createProcessWithFragmentParams(fragmentId, List(evaluatedparam.Parameter(paramName, "\"Tomasz\"")))
+
+    val processValidation = mockedProcessValidation(fragmentDefinition, defaultConfig)
+    val result            = processValidation.validate(processWithFragment)
+    print(result)
+    result.hasErrors shouldBe false
+    result.errors.invalidNodes shouldBe Symbol("empty")
+    result.errors.globalErrors shouldBe Symbol("empty")
+    result.saveAllowed shouldBe true
+  }
+
+  test("validates scenario with fragment parameters - with spel validation expression and invalid value") {
+    val fragmentId = "fragment1"
+    val paramName  = "name"
+
+    val configWithValidators: Config = defaultConfig
+
+    val fragmentDefinition: CanonicalProcess =
+      createFragmentDefinition(
+        fragmentId,
+        List(
+          FragmentParameter(
+            paramName,
+            FragmentClazzRef[String],
+            Some(ValidationExpression(s"#$paramName.length() < 7"))
+          )
+        )
+      )
+    val processWithFragment =
+      createProcessWithFragmentParams(fragmentId, List(evaluatedparam.Parameter(paramName, "\"Barabasz\"")))
+
+    val processValidation = mockedProcessValidation(fragmentDefinition, configWithValidators)
+    val result            = processValidation.validate(processWithFragment)
+    result.hasErrors shouldBe true
+    result.errors.globalErrors shouldBe empty
+    result.errors.invalidNodes.get("subIn") should matchPattern {
+      case Some(
+            List(
+              NodeValidationError(
+                "CustomParameterValidationError",
+                _,
+                _,
+                Some(paramName),
+                NodeValidationErrorType.SaveAllowed
+              )
             )
           ) =>
     }